// apps/web/src/components/editor/sidebar/scriptSideBar/types/ActionTypes.ts
export interface BaseAction {
  id: string;
  name: string;
  type: ActionType;
  enabled: boolean;
}
// ...
export interface MoveToAction extends BaseAction {
  type: ActionType.MOVE_TO;
  targetX: number;
  targetY: number;
  duration?: number;
  easeType?: EaseType;
}

export interface MoveByAction extends BaseAction {
  type: ActionType.MOVE_BY;
  deltaX: number;
  deltaY: number;
  duration?: number;
  easeType?: EaseType;
}

export interface RotateToAction extends BaseAction {
  type: ActionType.ROTATE_TO;
  targetRotation: number;
  duration?: number;
  easeType?: EaseType;
}

export interface RotateByAction extends BaseAction {
  type: ActionType.ROTATE_BY;
  deltaRotation: number;
  duration?: number;
  easeType?: EaseType;
}

export interface ScaleToAction extends BaseAction {
  type: ActionType.SCALE_TO;
  targetScaleX: number;
  targetScaleY: number;
  duration?: number;
  easeType?: EaseType;
}

export interface SetVisibleAction extends BaseAction {
  type: ActionType.SET_VISIBLE;
  visible: boolean;
}

export interface SetModulateAction extends BaseAction {
  type: ActionType.SET_MODULATE;
  color: string;
  alpha: number;
}

export interface SetPropertyAction extends BaseAction {
  type: ActionType.SET_PROPERTY;
  propertyPath: string;
  value: string;
  valueType: PropertyValueType;
}

export interface TweenPropertyAction extends BaseAction {
  type: ActionType.TWEEN_PROPERTY;
  propertyPath: string;
  targetValue: string;
  duration: number;
  easeType: EaseType;
  transitionType: TransitionType;
}

export interface EmitSignalAction extends BaseAction {
  type: ActionType.EMIT_SIGNAL;
  signalName: string;
  parameters: SignalParameter[];
}

export interface WaitAction extends BaseAction {
  type: ActionType.WAIT;
  duration: number;
  waitType: WaitType;
}

export interface ConditionAction extends BaseAction {
  type: ActionType.CONDITION;
  leftOperand: string;
  operator: ComparisonOperator;
  rightOperand: string;
  trueActions: string[]; // Action IDs
  falseActions: string[]; // Action IDs
}

export interface PrintAction extends BaseAction {
  type: ActionType.PRINT;
  message: string;
  includeNodeName: boolean;
}

export interface PlaySoundAction extends BaseAction {
  type: ActionType.PLAY_SOUND;
  soundPath: string;
  volume: number;
  pitch: number;
  loop: boolean;
}

export interface CustomCodeAction extends BaseAction {
  type: ActionType.CUSTOM_CODE;
  code: string;
}

// New Godot Action Interfaces

// Node2D Transform Actions
export interface TranslateAction extends BaseAction {
  type: ActionType.TRANSLATE;
  vector: { x: number; y: number };
}

export interface RotateAction extends BaseAction {
  type: ActionType.ROTATE;
  angle: number;
}

export interface LookAtAction extends BaseAction {
  type: ActionType.LOOK_AT;
  point: { x: number; y: number };
}

export interface GetAngleToAction extends BaseAction {
  type: ActionType.GET_ANGLE_TO;
  point: { x: number; y: number };
}

export interface ToLocalAction extends BaseAction {
  type: ActionType.TO_LOCAL;
  globalPoint: { x: number; y: number };
}

export interface ToGlobalAction extends BaseAction {
  type: ActionType.TO_GLOBAL;
  localPoint: { x: number; y: number };
}

// Node Core Actions
export interface AddChildAction extends BaseAction {
  type: ActionType.ADD_CHILD;
  childPath: string;
}

export interface RemoveChildAction extends BaseAction {
  type: ActionType.REMOVE_CHILD;
  childPath: string;
}

export interface GetNodeAction extends BaseAction {
  type: ActionType.GET_NODE;
  nodePath: string;
}

export interface SetProcessAction extends BaseAction {
  type: ActionType.SET_PROCESS;
  enabled: boolean;
}

export interface SetPhysicsProcessAction extends BaseAction {
  type: ActionType.SET_PHYSICS_PROCESS;
  enabled: boolean;
}

// Sprite2D Actions
export interface SetTextureAction extends BaseAction {
  type: ActionType.SET_TEXTURE;
  texturePath: string;
}

export interface SetRegionEnabledAction extends BaseAction {
  type: ActionType.SET_REGION_ENABLED;
  enabled: boolean;
}

export interface SetRegionRectAction extends BaseAction {
  type: ActionType.SET_REGION_RECT;
  x: number;
  y: number;
  width: number;
  height: number;
}

export interface SetFlipHAction extends BaseAction {
  type: ActionType.SET_FLIP_H;
  flip: boolean;
}

export interface SetFlipVAction extends BaseAction {
  type: ActionType.SET_FLIP_V;
  flip: boolean;
}

// AnimatedSprite2D Actions
export interface StopAnimationAction extends BaseAction {
  type: ActionType.STOP_ANIMATION;
}

export interface PauseAnimationAction extends BaseAction {
  type: ActionType.PAUSE_ANIMATION;
}

export interface SetAnimationAction extends BaseAction {
  type: ActionType.SET_ANIMATION;
  animationName: string;
}

export interface SetFrameAction extends BaseAction {
  type: ActionType.SET_FRAME;
  frame: number;
}

export interface SetSpeedScaleAction extends BaseAction {
  type: ActionType.SET_SPEED_SCALE;
  speedScale: number;
}

// RigidBody2D Actions
export interface ApplyImpulseAction extends BaseAction {
  type: ActionType.APPLY_IMPULSE;
  impulse: { x: number; y: number };
  position?: { x: number; y: number };
}

export interface ApplyForceAction extends BaseAction {
  type: ActionType.APPLY_FORCE;
  force: { x: number; y: number };
  position?: { x: number; y: number };
}

export interface SetGravityScaleAction extends BaseAction {
  type: ActionType.SET_GRAVITY_SCALE;
  scale: number;
}

export interface FreezeBodyAction extends BaseAction {
  type: ActionType.FREEZE_BODY;
}

export interface SetFreezeEnabledAction extends BaseAction {
  type: ActionType.SET_FREEZE_ENABLED;
  enabled: boolean;
}

export interface SetMassAction extends BaseAction {
  type: ActionType.SET_MASS;
  mass: number;
}

export interface SetLinearVelocityAction extends BaseAction {
  type: ActionType.SET_LINEAR_VELOCITY;
  velocity: { x: number; y: number };
}

export interface SetAngularVelocityAction extends BaseAction {
  type: ActionType.SET_ANGULAR_VELOCITY;
  velocity: number;
}

export interface SetLockRotationAction extends BaseAction {
  type: ActionType.SET_LOCK_ROTATION;
  locked: boolean;
}

// StaticBody2D Actions
export interface MoveAndCollideAction extends BaseAction {
  type: ActionType.MOVE_AND_COLLIDE;
  velocity: { x: number; y: number };
}

export interface SetConstantLinearVelocityAction extends BaseAction {
  type: ActionType.SET_CONSTANT_LINEAR_VELOCITY;
  velocity: { x: number; y: number };
}

export interface SetConstantAngularVelocityAction extends BaseAction {
  type: ActionType.SET_CONSTANT_ANGULAR_VELOCITY;
  velocity: number;
}

// Area2D Actions
export interface GetOverlappingBodiesAction extends BaseAction {
  type: ActionType.GET_OVERLAPPING_BODIES;
}

export interface GetOverlappingAreasAction extends BaseAction {
  type: ActionType.GET_OVERLAPPING_AREAS;
}

export interface OverlapsBodyAction extends BaseAction {
  type: ActionType.OVERLAPS_BODY;
  bodyPath: string;
}

export interface OverlapsAreaAction extends BaseAction {
  type: ActionType.OVERLAPS_AREA;
  areaPath: string;
}

export interface SetMonitoringAction extends BaseAction {
  type: ActionType.SET_MONITORING;
  enabled: boolean;
}

export interface SetMonitorableAction extends BaseAction {
  type: ActionType.SET_MONITORABLE;
  enabled: boolean;
}

export interface SetGravityAction extends BaseAction {
  type: ActionType.SET_GRAVITY;
  gravity: number;
}

export interface SetGravityVectorAction extends BaseAction {
  type: ActionType.SET_GRAVITY_VECTOR;
  vector: { x: number; y: number };
}

// CharacterBody2D Actions
export interface MoveAndSlideAction extends BaseAction {
  type: ActionType.MOVE_AND_SLIDE;
}

export interface IsOnFloorAction extends BaseAction {
  type: ActionType.IS_ON_FLOOR;
}

export interface IsOnCeilingAction extends BaseAction {
  type: ActionType.IS_ON_CEILING;
}

export interface IsOnWallAction extends BaseAction {
  type: ActionType.IS_ON_WALL;
}

export interface GetFloorNormalAction extends BaseAction {
  type: ActionType.GET_FLOOR_NORMAL;
}

export interface SetVelocityAction extends BaseAction {
  type: ActionType.SET_VELOCITY;
  velocity: { x: number; y: number };
}

export interface SetFloorMaxAngleAction extends BaseAction {
  type: ActionType.SET_FLOOR_MAX_ANGLE;
  angle: number;
}

export interface SetUpDirectionAction extends BaseAction {
  type: ActionType.SET_UP_DIRECTION;
  direction: { x: number; y: number };
}

// CollisionShape2D Actions
export interface SetShapeAction extends BaseAction {
  type: ActionType.SET_SHAPE;
  shapePath: string;
}

export interface SetCollisionDisabledAction extends BaseAction {
  type: ActionType.SET_COLLISION_DISABLED;
  disabled: boolean;
}

export interface SetOneWayCollisionAction extends BaseAction {
  type: ActionType.SET_ONE_WAY_COLLISION;
  enabled: boolean;
}

// CollisionPolygon2D Actions
export interface SetPolygonAction extends BaseAction {
  type: ActionType.SET_POLYGON;
  points: Array<{ x: number; y: number }>;
}

export interface SetBuildModeAction extends BaseAction {
  type: ActionType.SET_BUILD_MODE;
  mode: BuildMode;
}

export enum BuildMode {
  SOLIDS = 'solids',
  SEGMENTS = 'segments'
}

// Union type for all actions
export type Action = 
  | MoveToAction 
  | MoveByAction 
  | RotateToAction 
  | RotateByAction 
  | ScaleToAction
  | SetVisibleAction
  | SetModulateAction
  | SetPropertyAction
  | TweenPropertyAction
  | EmitSignalAction
  | WaitAction
  | ConditionAction
  | PrintAction
  | PlaySoundAction
  | CustomCodeAction
  // New Godot Actions
  | TranslateAction
  | RotateAction
  | LookAtAction
  | GetAngleToAction
  | ToLocalAction
  | ToGlobalAction
  | AddChildAction
  | RemoveChildAction
  | GetNodeAction
  | SetProcessAction
  | SetPhysicsProcessAction
  | SetTextureAction
  | SetRegionEnabledAction
  | SetRegionRectAction
  | SetFlipHAction
  | SetFlipVAction
  | StopAnimationAction
  | PauseAnimationAction
  | SetAnimationAction
  | SetFrameAction
  | SetSpeedScaleAction
  | ApplyImpulseAction
  | ApplyForceAction
  | SetGravityScaleAction
  | FreezeBodyAction
  | SetFreezeEnabledAction
  | SetMassAction
  | SetLinearVelocityAction
  | SetAngularVelocityAction
  | SetLockRotationAction
  | MoveAndCollideAction
  | SetConstantLinearVelocityAction
  | SetConstantAngularVelocityAction
  | GetOverlappingBodiesAction
  | GetOverlappingAreasAction
  | OverlapsBodyAction
  | OverlapsAreaAction
  | SetMonitoringAction
  | SetMonitorableAction
  | SetGravityAction
  | SetGravityVectorAction
  | MoveAndSlideAction
  | IsOnFloorAction
  | IsOnCeilingAction
  | IsOnWallAction
  | GetFloorNormalAction
  | SetVelocityAction
  | SetFloorMaxAngleAction
  | SetUpDirectionAction
  | SetShapeAction
  | SetCollisionDisabledAction
  | SetOneWayCollisionAction
  | SetPolygonAction
  | SetBuildModeAction;

// Enums for options
export enum EaseType {
  LINEAR = 'linear',
  EASE_IN = 'ease_in',
  EASE_OUT = 'ease_out',
  EASE_IN_OUT = 'ease_in_out',
  EASE_OUT_IN = 'ease_out_in'
}

export enum TransitionType {
  SINE = 'sine',
  QUINT = 'quint',
  QUART = 'quart',
  QUAD = 'quad',
  EXPO = 'expo',
  ELASTIC = 'elastic',
  CUBIC = 'cubic',
  CIRC = 'circ',
  BOUNCE = 'bounce',
  BACK = 'back'
}

export enum PropertyValueType {
  STRING = 'string',
  NUMBER = 'number',
  BOOLEAN = 'boolean',
  VECTOR2 = 'vector2',
  COLOR = 'color'
}

export enum WaitType {
  SECONDS = 'seconds',
  FRAMES = 'frames',
  SIGNAL = 'signal'
}

export enum ComparisonOperator {
  EQUALS = '==',
  NOT_EQUALS = '!=',
  GREATER = '>',
  LESS = '<',
  GREATER_EQUAL = '>=',
  LESS_EQUAL = '<=',
  AND = 'and',
  OR = 'or'
}

export interface SignalParameter {
  name: string;
  value: string;
  type: PropertyValueType;
}